Add citext to the list of default column types
As part of Rails 4.2, citext columns in PostgreSQL databases are
recognized as a type distinct from `:text` and `:string` – `:citext` –
when previously they were an unrecognized type and hence treated as
`:string`s.

Since these columns are a textual type, and so can evaluate to blank,
they should be in the default column types so they are not missed when
using `nilify_blanks` on a model with citext columns.